Swap the Transition and Tween names
Tween<T> is the spec now - duration, easing, interpolation, Copy,
stateless. Transition<H, T> is the running interpolation - base,
target, elapsed, the write pointer - kept in a TransitionTable<H> per
node and field. The files swap to match: tween.rs holds the spec,
transition.rs the runner.

insert_tween -> insert_transition, TweenTable -> TransitionTable,
Fynix::tween_len -> transition_len, Records::tweens ->
Records::transitions.
