+++ crates/fynix_macros/src/element.rs
@@ -264,9 +264,9 @@ pub fn expand(
     let build_hook = opts.build.map(|build_fn| {
         quote! {
             {
-                let (__tweens, __store) = records.build_parts();
+                let (__transitions, __store) = records.build_parts();
                 let mut __draw = #root::ui::Build::new(
-                    world, node, __tweens, __store, theme,
+                    world, node, __transitions, __store, theme,
                 );
                 (#build_fn)(self, &mut __draw);
             }
